get_process_watcher - stop saving state, for now
The purpose of the state.py file is to allow re-acquisition of running
cluster jobs after pypeflow dies and is restarted. But we have not
implemented that yet anyway. When we do, we'll figure out a way to
ensure that state.py is written on exit.

By neglecting it, we can more easily support the new blocking
process-watcher: #31